NominationInfo in one
