+++ packages/app/src/components/nominationInfo/applicationDetail.js
@@ -9,7 +9,7 @@ function ApplicationDetail(props) {
           </div>
           <div className={[styles.content, (props.gridContent && styles["grid-container"])].join(" ")}>
             {
-              props.fields.map((obj) => (<div key={obj.label} className={obj.label === "" && styles.mobileHide}>
+              props.fields.map((obj) => (<div key={obj.label} className={obj.label === "" ? styles.mobileHide: ""}>
                 <label>{obj.label}</label>
                 <span>{String(obj.value)}</span>
               </div>))
