+++ packages/app/src/components/nominationInfo/applicationDetail.test.js
